backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
python-dotenv==1.0.1
